fix photo display issue
